+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/Track.kt
@@ -16,7 +16,7 @@ interface Track : Serializable {
 
     var title: String
 
-    var last_chapter_read: Float
+    var last_chapter_read: Double
 
     var total_chapters: Int
