+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangaupdates/dto/ListItem.kt
@@ -17,6 +17,6 @@ data class ListItem(
 fun ListItem.copyTo(track: Track): Track {
     return track.apply {
         this.status = listId ?: READING_LIST
-        this.last_chapter_read = this@copyTo.status?.chapter?.toFloat() ?: 0f
+        this.last_chapter_read = this@copyTo.status?.chapter?.toDouble() ?: 0.0
     }
 }
