+++ app/src/main/java/eu/kanade/tachiyomi/data/track/myanimelist/MyAnimeListApi.kt
@@ -16,6 +16,7 @@ import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.boolean
 import kotlinx.serialization.json.contentOrNull
+import kotlinx.serialization.json.double
 import kotlinx.serialization.json.doubleOrNull
 import kotlinx.serialization.json.float
 import kotlinx.serialization.json.floatOrNull
@@ -250,7 +251,7 @@ class MyAnimeListApi(
         return track.apply {
             val isRereading = obj["is_rereading"]!!.jsonPrimitive.boolean
             status = if (isRereading) MyAnimeList.REREADING else getStatus(obj["status"]?.jsonPrimitive?.content)
-            last_chapter_read = obj["num_chapters_read"]!!.jsonPrimitive.float
+            last_chapter_read = obj["num_chapters_read"]!!.jsonPrimitive.double
             score = obj["score"]!!.jsonPrimitive.int.toDouble()
             obj["start_date"]?.let {
                 started_reading_date = parseDate(it.jsonPrimitive.content)
